Predict typing indicator
It greatly annoys me in the rare instance I do play.
